requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.2.1
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.1